// proyect-web/src/services/api/api.enum.ts
/**
 * Enum for the different game phases
 */
export enum GamePhaseEnum {
    RECRUITMENT = 'recruitment',
	STARTING = 'starting',
	WORDS = 'words',
	FINISHED = 'finished',
	VOTING = 'voting',
	VOTING_RESULTS = 'votingResults'
}

/**
 * Rol of the player at the party
 */
export enum PlayerRolEnum {
	MASTER = 'master',
    PLAYER = 'player'
}